api/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
click==8.0.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
mux-python @ git+https://github.com/muxinc/mux-python.git@c2df765035d457dd98ae69c3466e48f22964c43a
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
